header-includes-template.md (from .html)
From pandoc docs:

Raw content to include in the document’s header may be specified
using header-includes; however, it is important to mark up this
content as raw code for a particular output format, using the
raw_attribute extension), or it will be interpreted as markdown.
