+++ manubot/process/metadata.py
@@ -1,5 +1,5 @@
 """
-Tools for manuscript thumbnail detection and processing.
+Tools for manuscript metadata processing including thumbnail detection and processing.
 """
 import functools
 import logging
@@ -7,6 +7,21 @@ import pathlib
 import subprocess
 from typing import Optional
 from urllib.parse import urljoin
+
+
+def get_header_includes(variables: dict) -> str:
+    """
+    Render `header-includes-template.md` using information from `variables`.
+    """
+    from .util import template_with_jinja2
+
+    path = pathlib.Path(__file__).parent.joinpath("header-includes-template.md")
+    try:
+        template = path.read_text(encoding="utf-8-sig")
+        return template_with_jinja2(template, variables)
+    except Exception:
+        logging.exception(f"Error generating header-includes.")
+        return ""
 
 
 def get_thumbnail_url(thumbnail=None):
