+++ manubot/process/tests/test_util.py
@@ -1,9 +1,8 @@
-import copy
 import pathlib
 
 import pytest
 
-from ..util import add_author_affiliations, get_header_includes, read_variable_files
+from ..util import add_author_affiliations, read_variable_files
 
 directory = pathlib.Path(__file__).parent.resolve()
 
@@ -80,32 +79,3 @@ def test_add_author_affiliations():
     ]
     assert authors[0]["affiliation_numbers"] == [1, 2]
     assert authors[1]["affiliation_numbers"] == [2, 3]
-
-
-def test_get_header_includes_description_abstract():
-    # test that abstract and description set different fields if both supplied
-    variables = {
-        "pandoc": {"abstract": "value for abstract"},
-        "manubot": {"description": "value for description"},
-    }
-    header_includes = get_header_includes(copy.deepcopy(variables))
-    assert (
-        '<meta name="citation_abstract" content="value for abstract" />'
-        in header_includes
-    )
-    assert (
-        '<meta name="description" content="value for description" />' in header_includes
-    )
-    # test that abstract is used for description if description is not defined
-    del variables["manubot"]["description"]
-    header_includes = get_header_includes(copy.deepcopy(variables))
-    assert (
-        '<meta name="citation_abstract" content="value for abstract" />'
-        in header_includes
-    )
-    assert '<meta name="description" content="value for abstract" />' in header_includes
-    # test that if neither abstract nor description is defined, neither are inserted
-    del variables["pandoc"]["abstract"]
-    header_includes = get_header_includes(copy.deepcopy(variables))
-    assert 'meta name="citation_abstract"' not in header_includes
-    assert 'meta name="description"' not in header_includes
